+++ backend/handlers/signup_test.go
@@ -1,9 +1,9 @@
-package routes_test
+package handlers_test
 
 import (
+	"go_notion/backend/handlers"
 	"go_notion/backend/mocks"
 	"go_notion/backend/router"
-	"go_notion/backend/routes"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -61,7 +61,7 @@ func TestSignUp(t *testing.T) {
 			}
 
 			tokenGenerator := &mocks.TokenGeneratorMock{}
-			signUp, err := routes.NewSignUpHandler(mock, tokenGenerator)
+			signUp, err := handlers.NewSignUpHandler(mock, tokenGenerator)
 			if err != nil {
 				t.Fatal(err)
 			}
